job/job/pyproject.toml
[project]
name = "job"
version = "0.1.0"
description = "AI-powered resume analysis and job recommendations from LinkedIn and Naukri"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
keywords = [
    "resume",
    "job-search",
    "openai",
    "streamlit",
    "apify",
    "linkedin",
    "naukri",
    "mcp",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]
dependencies = [
    "apify-client>=1.10.0",
    "fastmcp>=2.0.0",
    "mcp[cli]>=1.9.2",
    "openai>=1.84.0",
    "pymupdf>=1.26.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
